Refactor Mat4 and Quat to use scalar/sse module pattern
- Create katla_math/src/scalar/ with scalar implementations
  - Add vec2.rs, vec3.rs, vec4.rs, mat4.rs, quat.rs
- Create katla_math/src/sse/ with SSE-accelerated implementations
  - Add vec4.rs with SIMD-optimized Vec4 operations
  - Add quat.rs with __m128-based Quat using SSE for dot, normalize, slerp
  - Add mat4.rs with SIMD-optimized extract_row and transpose
- Convert mat4.rs and quat.rs to re-export modules selecting implementation
  based on target architecture (x86/x86_64 uses SSE, others use scalar)
- Extract SSE shuffle bit patterns to named constants for readability
- Change Quat::to_mat3 and to_euler to take self by value (Copy type convention)
- Fix clippy warnings in katla_vulkan (unused import, mut, lifetime elision)
- Make Model::new_from_gltf_with_ptr pub(crate) to avoid unsafe ptr in public API

+++ katla_vulkan/src/vulkan/context.rs
@@ -1040,7 +1040,7 @@ unsafe fn pick_physical_device_headless(instance: &Instance) -> Option<vk::Physi
             _ => {}
         }
 
-        score += properties.limits.max_image_dimension2_d as u32;
+        score += properties.limits.max_image_dimension2_d;
         score
     });
